feat(T1c): PhysBERT VENCE a fusao (p=0,0062), e dois erros meus na mesma corrida
    variante              fusao   +PhiRank        d   disc   p(k=10)   veredito
    minilm (controle)    0,1576    0,1483  -0,0093    229   0,14584   empate
    phys (physbert)      0,1576    0,1666  +0,0090    237   0,00625   VENCE
    gte (gte-base)            -         -        -      -         -   FALHOU

A predicao do T1b se confirmou. O reranqueador de base DIFERENTE bate a fusao com
p=0,00625, abaixo do limiar de Bonferroni de 0,025 registrado antes de medir. E a
primeira vez que a composicao inteira funciona: recall@10 de 0,2890 contra 0,2675
da fusao, acerto@1 no grupo de 0,566 contra 0,498 do controle.

O controle replicou o empate do T1b com o dobro das consultas (p=0,146 contra
0,118 antes, 229 discordantes contra 105). Isso e o que mostra que a diferenca e
da base e nao do protocolo.

## Erro 1 — o gte morreu por bug nosso, nao dele

    ValueError: Attempting to unscale FP16 gradients.

`thenlper/gte-base` guarda os pesos em fp16, o transformers novo carrega no dtype
do checkpoint por padrao, e o GradScaler recusa desescalar gradientes fp16 — o AMP
exige pesos-mestres em fp32. Consertado com `dtype=torch.float32` explicito.

Ficou invisivel porque MiniLM e PhysBERT sao fp32. Qualquer base fp16 quebraria, e
o custo foi um braco inteiro do experimento.

## Erro 2 — o meu pre-registro confundia "perdeu" com "nao rodou"

Com o gte ausente, o script imprimiu "o mecanismo e CONHECIMENTO DE DOMINIO, nao
diversidade" — leitura que EXIGE o gte ter produzido um numero e perdido. A logica
olhava so `venceram`, nunca `falhas`. O pre-registro existia para impedir
exatamente esse tipo de conclusao, e a implementacao dele tinha o buraco.

Agora qualquer braco ausente torna o mecanismo INCONCLUSIVO, e ha um teste que
EXECUTA a logica nos quatro desfechos em vez de procurar texto.

**O que esta estabelecido:** uma base diferente vence. **O que segue aberto:** qual
propriedade da base faz isso — dominio ou capacidade.

## Vazao medida contra a minha estimativa

PhysBERT treinou a 35,4 exemplos/s com 2.332 MB de VRAM. Eu estimei 14-24/s —
pessimista por ~2x. As avaliacoes cairam na faixa (13 min o de 23 M, ~40 min o de
109 M). E 2,3 GB num cartao de 16 GB dizem que `--grupos 2` subutiliza a placa;
aumentar seria mais rapido e quebraria a comparabilidade com o controle, entao
fica.

`SO = {"gte"}` na celula existe para COMPLETAR o braco que falhou por bug de
infraestrutura, nao para escolher bracos depois de ver resultados — e o mecanismo
segue inconclusivo enquanto ele nao tiver numero.

503 testes passam, 9 pulam.

+++ src/phifm/training/rerank.py
@@ -157,8 +157,25 @@ class TreinadorRank:
         self.dev = escolher_dispositivo(cfg.dispositivo)
         self.tok = AutoTokenizer.from_pretrained(cfg.base)
         atencao = "sdpa" if self.dev.type == "cuda" else "eager"
+        # ⚠️ `dtype` EXPLÍCITO em fp32, e não o do checkpoint.
+        #
+        # Medido em 2026-08-31, no T1c: `thenlper/gte-base` guarda os pesos em fp16.
+        # O `transformers` novo carrega no dtype do checkpoint por padrão, e aí o
+        # `GradScaler` morre no primeiro passo:
+        #
+        #     ValueError: Attempting to unscale FP16 gradients.
+        #
+        # O AMP exige pesos-mestres em fp32 — é ele que faz a passagem em fp16, e o
+        # `GradScaler` que desescala os gradientes de volta. Um modelo já em fp16 não
+        # tem para onde desescalar.
+        #
+        # A base do controle (`all-MiniLM-L6-v2`) e o PhysBERT são fp32, então isto
+        # ficou invisível até a primeira base fp16 entrar. Não é propriedade do gte:
+        # qualquer checkpoint fp16 quebraria, e o custo foi um braço inteiro do
+        # experimento.
         self.mod = AutoModelForSequenceClassification.from_pretrained(
-            cfg.base, num_labels=1, attn_implementation=atencao).to(self.dev)
+            cfg.base, num_labels=1, attn_implementation=atencao,
+            dtype=torch.float32).to(self.dev)
         if cfg.checkpointing:
             self.mod.gradient_checkpointing_enable()
             self.mod.config.use_cache = False
